refactor: renames and moves di container files to different location
