+++ apps/deploy-web/src/context/ServicesProvider/ServicesContext.ts
@@ -1,6 +1,6 @@
 import { createContext } from "react";
 
-import { services as rootContainer } from "@src/services/http/http-browser.service";
+import { services as rootContainer } from "@src/services/app-di-container/browser-di-container";
 
 export type RootContainer = typeof rootContainer;
