+++ apps/deploy-web/src/lib/nextjs/defineApiHandler/defineApiHandler.spec.ts
@@ -3,7 +3,7 @@ import { mock } from "jest-mock-extended";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { z } from "zod";
 
-import { services } from "@src/services/http/http-server.service";
+import { services } from "@src/services/app-di-container/server-di-container.service";
 import type { NextApiRequestWithServices } from "./defineApiHandler";
 import { defineApiHandler, REQ_SERVICES_KEY } from "./defineApiHandler";
